+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * :sparkles: Arrayx add orEmpty method
 * :hammer: The primary parameters of all Arrayx methods are now Nullable
 * :sparkles: Filex add lengthRecursively(File\[\]) and lengthRecursively(Iterable<File>) method
+* :sparkles: IOStreamx copyTo method add CopyListener argument
 
 ## 0.6.5
 * :hammer: String orContains and andContains change to extension method